function relocation
move functions from distparameters to GISpy and another new file WeirdFn
and update setup file
