+++ Hapi/__init__.py
@@ -39,6 +39,7 @@ import Routing
 import RUN
 import StatisticalTools
 import Wrapper
+import Weird_Fn
 
 
 # module level doc-string
